fix: support PreparedStatement.setBlob(1, Blob) and PreparedStatement.setClob(1, Clob) for lobs that return -1 for length
Historically, Hibernate was creating such clobs and it relied on the fact that pgjdbc
did not use the length in case it was -1.

Now the negative length is treated as if it was infinite (no eplicit limit)
The length of 0 means "empty lob".

Fixes https://github.com/pgjdbc/pgjdbc/issues/3134