Force getMetaData() via P & D messages even in simple mode
PreparedStatement.getMetaData() can be called before binding of
parameters, so the "?" in the query cannot be expanded, the SQL would be
sent to database with "?" unexpanded, this will cause syntax error of
course.

To let this work we have to force the method run via the parse &
describe libpq messages.