Introduce optional global TypeInfoCache
In the current implementation, `TypeInfoCache` is a per-connection
cache. This means, over time, that the same type information is
retrieved from the database copious times. For some DB systems, this
has a negative cost/performance impact.

Introduce an optional global `TypeInfoCache`. This is useful for
cases where a single DB system is used in Production for all
connections and there's no need for each connection to re-retrieve
type info.

- Extract the storage portion of `TypeInfoCache` into a new class,
`TypeInfoCacheStorage`
- Add new property, `GLOBAL_TYPE_INFO_CACHE` to enable a global cache
- When enabled, the constant `TypeInfoCacheStorage GLOBAL_INSTANCE` is
always used. Over the course of running a Production instance, this global
cache should not need to query the database for type info.