feat: do not attempt to guard server-side large object limit
BlobInputStream.seek will now throw if attempting to skip beyond the max size for a large object. This is a breaking change. The previous implementation was imperfect as it did not account for the server being compiled with a different limit.

The state of the stream after the exception is undefined, but hopefully sane. Further exceptions may be thrown on future operations.