WIP: implement direct SSL ALPN connections
Not sure when, but recent versions of openssl have deprecated PKCS-8 keys
Removed the test for PKCS-8 keys
Calling TcpNoDelay on an SSL connection that did not complete the SSL negotiation
results in an Invalid argument exception instead of a broken pipe exception
This caused the tests to fail. Changed the test to allow for this exception.

Removed test for pk8

Checkstyle

rename LazyManagerKey test to PKCS12KeyManagerTest
removed tests for LazyKeyManager

Add documentation
add tests/