refactor: factor out duplicated .getBytes() when converting date/time to Date/Time/Timestamp
At best, we should avoid `byte[] -> String -> byte[]` conversions
when dealing with datetime, however, currently PgCallableStatement uses
getObject eagerly, so we lose the underlying wire format byte[] early.

That is why we have to use .toString() when converting to date.
Ideally, we should keep the source byte[] data and convert it on request.