test: simplify TestUtil.openDB, add tests with various assumeMinServerVersion values
Previously, the code considered only a few of PGProperty names when creating a database connection.
So there was no possibility to set the rest of the properties by passing them with build.properties or -D...

After the change, the code uses all the matching properties from System.properties,
so we can increase test coverage.

For instance, the change adds tests with assumeMinServerVersion setting.