Merge commit from fork
* Fix channel binding required handling to reject non-SASL authentication

If channel binding is set to required but SASL authentication has either not completed
or a different form of authentication is being performed, then we now throw an error
as the user has explicitly requested channel binding but it is not possible.

Previously we ignored the channel binding mode of required for non-SASL authentication.

* Add ChannelBindingRequiredTest

* prepare for release of 42.7.7

---------

Co-authored-by: Dave Cramer <davecramer@gmail.com>