fix: remove excessive ReentrantLock.lock usages
In some of the cases (private/protected/internal API) we know we already
hold the lock, so there's no much sense aquiring it again.

For LruCache, we know we do not perform IO operations, so it is safe to use
synchronized as it won't cause Loom virtual thread pinning.

See https://github.com/pgjdbc/pgjdbc/issues/3693