feat: support java.time.Instant for ResultSet.getObject(..., Class) and PreparedStatement.set(..., Object)
Previously, the users had to round-trip to OffsetDateTime.class which had several drawbacks:
1) It was not convenient when the application already had Instant at hand
2) Instant has a closer mapping to the underlying PostgreSQL timestamptz type (== seconds from epoch),
  so it is a bit faster if the user does not need date/time components.

Only timestamptz and timetz are supported for Instant.
timestamp in the database does not properly represent a moment in time,
so we fail fast so the application developer could either alter the type of the column
or use a different type in Java (e.g. LocalDateTime)

Fixes https://github.com/pgjdbc/pgjdbc/issues/1325