feat: support java.time.ZonedDateTime for ResultSet.getObject(..., Class) and PreparedStatement.set(..., Object)
Previously, the users had to round-trip to OffsetDateTime.class which had several drawbacks:
1) It was not convenient when the application already had Instant at hand

In fact, PostgreSQL did already support TIMESTAMP '... ZoneName` literals,
so if we convert ZonedDateTime to OffsetDateTime when binding the parameter,
then we keep pretty much the same behaviour as the database.

Only timestamptz and timetz are supported for ZonedDateTime.
timestamp in the database does not properly represent a moment in time,
so we fail fast so the application developer could either alter the type of the column
or use a different type in Java (e.g. LocalDateTime)

Fixes https://github.com/pgjdbc/pgjdbc/issues/1325