chore: use bump org.apache.bcel:bcel test dependency in testCompileClasspath as well
Previously, bcel was updated only in testRuntimeClasspath, so it triggered alarms like "old BCEL used in compilation".

Now we use an updated bcel for both compile and the runtime, and we use
constraints rather than direct dependency, so bcel no longer participates
in the classpath if third-party libraries move away from bcel.