Fix: BaseDataSource.getProperty mirrors BaseDataSource.setProperty.
The setProperty method of BaseDataSource allows all properties to be set.
The getProperty method does not allow the retrieval of: PG_HOST, PG_PORT, PG_DBNAME, USER, and PASSWORD.
All properties that can be set should also be readable.

Create unit tests to verify that getting and setting properties works as expected.

Fixes: https://github.com/pgjdbc/pgjdbc/issues/3761