fix: improve concurrency for prepareStatement
Previously, prepareStatement synchronized on a per-connection statement cache
for parsing the statement.
Technically speaking, it is not the best idea to share the same connections
across several execution threads, however, the driver does not need to heavily
synchronize when parsing the SQL.

Issue 3835 shows a case it might result in a carrier thread pinning, so
it still makes sense to move .create() outside of the synchronization.

The impact/improvement on the applications should be minimal provided
the application do not share a connection across concurrent execution threads.

Fixes https://github.com/pgjdbc/pgjdbc/issues/3835